requirements.txt
asgiref==3.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.18
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.21.5
execnet==1.9.0
idna==3.4
importlib-metadata==4.8.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.8
pytz==2022.7.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==1.2.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.6.0
